add support for DBZ postgresql CDC
initial support for postgres is based on the observation that DBZ basically produces similar CDC events for mysql and postgres. thus both can for now be supported based on the same code as written for mysql. refactorings have been made to avoid redundancy and allow for specific sub-classes of RdbmsHandler