add simple batch support for sink record processing (#51)
There is a new configuration parameter mongodb.max.batch.size=N to specify
the batching behaviour. For reasons of backwards compatibility the default value
of N=0 means that no batching is applied. For values of N>=1 the write
behaviour towards the configured MongoDB instance guarantees that
at most N records are written within a single bulkwrite operation.

This resolves #49