Configuration simplifications
Changed the focus of the configuration to be topic based.
Simplified the override process to be in the form:

`topic.override.<topicName>.<configuration>=<value>`

All configuration and overrides are now validated automatically.

Removed the following configurations: `MONGODB_COLLECTIONS_CONF` and `MONGODB_DOCUMENT_ID_STRATEGIES_CONF`
Removed the `mongodb.` prefix to configurations.
PostProcessor, updated the abstract class to take a topic based configuration and changed the design to be immutable.
Moved code to the `com.mongodb.kafka.connect.sink` namespace.

KAFKA-14
KAFKA-15