Allow the Sink connector to use multiple tasks
Kafka Connect provides built-in support for parallelism and scalable data copying
by assigning topic partitions to tasks. This allows for parallelism at the cost of
sequentially processing the data.

When using multiple tasks, data will be processeed out of order to the order of
the topic. Each task will be assigned partitions from a topic and these will be
processed independently of the other partitions.

KAFKA-92