Added Dynamic Topic Mapping for the Source
Added a new interface `TopicMapper` with a `getTopic` method. Implementations can take the
Change Stream Document to determine the correct `topic` to publish the data to.

The Source connector includes a single implementation: `DefaultTopicMapper`.
The `DefaultTopicMapper` uses the 'ns' field to determine the topic to publish to. The following
configuration options can help configure topic mapping:

  * `topic.prefix` - Add a prefix to the topic eg: `<prefix>.<databaseName>.<collectionName>`
  * `topic.suffix` - Add a suffix to the topic eg: `<databaseName>.<collectionName>.<suffix>`
  * `topic.namespace.map` - Set a json map to maps change stream document namespaces to topics:
    For example: `{\"db\": \"dbTopic\", \"db.coll\": \"dbCollTopic\"}` will map all Change Stream
    Documents from the `db` database to `dbTopic.<collectionName>` apart from any documents
    from the `db.coll` namespace which map to the `dbCollTopic` topic.
  * `topic.mapper` - use an alternative mapping class, users can implement their own
    `TopicMapper` class to handle the mapping logic.

KAFKA-185