Improved the usability of the Source connector heartbeat mechanism
Added the `heartbeat.bootstrap.servers` configuration so that
the HeartbeatManager can also consume heartbeat messages and
automatically update the offset.

Allowing the change stream postBatchResumeToken to be automatically
generated and consumed. On restart the latest seen offset (resumeToken)
will be used as the starting point for the new change stream cursor.

KAFKA-1767