this is a bug where copy existing does not turn off even after the task switched to change stream
the isCopying flag tracks wether the connector task is copying collection data or tracking change stream. This flag needs to be persisted, so that if the connector restarts later it does no start copying the collection again. There was a if statement which persisted this flag only when iscopying is true which is wrong, as the iscopying flag needs to be persisted in offsets after it has turned false when it has finished copying